+++ src/main/java/io/github/astrapi69/time/formatter/DateTimeFormatterFactory.java
@@ -27,7 +27,7 @@ package io.github.astrapi69.time.formatter;
 import java.time.format.DateTimeFormatter;
 import java.util.Locale;
 
-import io.github.astrapi69.time.enums.DatePattern;
+import io.github.astrapi69.time.enumtype.DatePattern;
 
 /**
  * The factory class {@link DateTimeFormatterFactory} provides methods for create