+++ src/test/java/io/github/astrapi69/time/LocalDateExtensionsTest.java
@@ -30,10 +30,10 @@ import java.sql.Date;
 import java.time.LocalDate;
 import java.util.Locale;
 
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.meanbean.test.BeanTester;
 
-import io.github.astrapi69.time.enums.DatePattern;
+import io.github.astrapi69.time.enumtype.DatePattern;
 
 /**
  * The unit test class for the class {@link LocalDateExtensions}
@@ -64,36 +64,43 @@ public class LocalDateExtensionsTest
 	 * Test method for {@link LocalDateExtensions#toDateString(LocalDate, String)}
 	 */
 	@Test
-	@Disabled
 	void testToDateString()
 	{
 		String actual;
 		String expected;
 		LocalDate input;
+		Locale defaultLocale;
+		// new scenario
+		defaultLocale = Locale.getDefault();
 		// new scenario
 		input = LocalDate.of(2000, 9, 1);
 		actual = LocalDateExtensions.toDateString(input, DatePattern.DDMMM);
-		expected = "01Sep.";
+		expected = LocalDateExtensions.toDateString(input, DatePattern.DDMMM.getPattern(),
+			defaultLocale);
 		assertEquals(expected, actual);
 
 		// new scenario
 		actual = LocalDateExtensions.toDateString(input, DatePattern.DDMM);
-		expected = "0109";
+		expected = LocalDateExtensions.toDateString(input, DatePattern.DDMM.getPattern(),
+			defaultLocale);
 		assertEquals(expected, actual);
 
 		// new scenario
 		actual = LocalDateExtensions.toDateString(input, DatePattern.MMDD);
-		expected = "0901";
+		expected = LocalDateExtensions.toDateString(input, DatePattern.MMDD.getPattern(),
+			defaultLocale);
 		assertEquals(expected, actual);
 
 		// new scenario
 		actual = LocalDateExtensions.toDateString(input, DatePattern.MMYY);
-		expected = "0900";
+		expected = LocalDateExtensions.toDateString(input, DatePattern.MMYY.getPattern(),
+			defaultLocale);
 		assertEquals(expected, actual);
 
 		// new scenario
 		actual = LocalDateExtensions.toDateString(input, DatePattern.YYYY);
-		expected = "2000";
+		expected = LocalDateExtensions.toDateString(input, DatePattern.YYYY.getPattern(),
+			defaultLocale);
 		assertEquals(expected, actual);
 	}
 
@@ -101,7 +108,6 @@ public class LocalDateExtensionsTest
 	 * Test method for {@link LocalDateExtensions#toDateString(LocalDate, String)}
 	 */
 	@Test
-	@Disabled
 	void testToDateStringWithLocale()
 	{
 		String actual;
@@ -112,7 +118,6 @@ public class LocalDateExtensionsTest
 		locale = Locale.GERMANY;
 		input = LocalDate.of(2000, 9, 1);
 		actual = LocalDateExtensions.toDateString(input, DatePattern.DDMMM.getPattern(), locale);
-		System.err.println(actual);
 		expected = "01Sep.";
 		assertEquals(expected, actual);
 
@@ -135,6 +140,16 @@ public class LocalDateExtensionsTest
 		actual = LocalDateExtensions.toDateString(input, DatePattern.YYYY.getPattern(), locale);
 		expected = "2000";
 		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link LocalDateExtensions}
+	 */
+	@Test
+	public void testWithBeanTester()
+	{
+		final BeanTester beanTester = new BeanTester();
+		beanTester.testBean(LocalDateExtensions.class);
 	}
 
 }