+++ src/test/java/io/github/astrapi69/time/convert/ZonedDateTimeExtensionsTest.java
@@ -22,11 +22,13 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.time;
+package io.github.astrapi69.time.convert;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
 import java.util.Date;
 
 import org.junit.jupiter.api.Test;
@@ -35,34 +37,66 @@ import org.meanbean.test.BeanTester;
 import de.alpharogroup.date.CreateDateExtensions;
 
 /**
- * The unit test class for the class {@link LocalDateTimeExtensions}
+ * The unit test class for the class {@link ZonedDateTimeExtensions}
  */
-public class LocalDateTimeExtensionsTest
+public class ZonedDateTimeExtensionsTest
 {
 
 	/**
-	 * Test method for {@link LocalDateTimeExtensions#toDate(LocalDateTime)}
+	 * Test method for {@link ZonedDateTimeExtensions#toDate(ZonedDateTime)}
 	 */
 	@Test
-	void toDate()
+	void testToDate()
 	{
 		Date actual;
 		Date expected;
-		LocalDateTime input;
+		ZonedDateTime input;
+		ZoneId zoneId;
+		LocalDateTime localDateTime;
 
-		input = LocalDateTime.of(2000, 9, 1, 0, 0, 0);
-		actual = LocalDateTimeExtensions.toDate(input);
+		localDateTime = LocalDateTime.of(2000, 9, 1, 0, 0, 0);
+		zoneId = ZoneId.systemDefault();
+
+		input = ZonedDateTime.of(localDateTime, zoneId);
+		actual = ZonedDateTimeExtensions.toDate(input);
 		expected = CreateDateExtensions.newDate(2000, 9, 1);
+		if (!actual.equals(expected))
+		{
+			java.util.logging.Logger logger = java.util.logging.Logger
+				.getLogger(this.getClass().getName());
+			logger.severe("  actual:" + actual + "\n" + "expected:" + expected);
+		}
+		assertEquals(expected, actual);
+	}
+
+
+	/**
+	 * Test method for {@link ZonedDateTimeExtensions#toZonedDateTime(LocalDateTime, ZoneId)}
+	 */
+	@Test
+	void testToZonedDateTime()
+	{
+		ZonedDateTime actual;
+		ZonedDateTime expected;
+		ZoneId zoneId;
+		LocalDateTime localDateTime;
+
+		localDateTime = LocalDateTime.of(2000, 9, 1, 0, 0, 0);
+		zoneId = ZoneId.of("Europe/Paris");
+
+		actual = ZonedDateTimeExtensions.toZonedDateTime(localDateTime, zoneId);
+		expected = ZonedDateTime.parse("2000-09-01T00:00+02:00[Europe/Paris]");
 		assertEquals(expected, actual);
 	}
 
 	/**
-	 * Test method for {@link LocalDateTimeExtensions}
+	 * Test method for {@link ZonedDateTimeExtensions}
 	 */
 	@Test
 	public void testWithBeanTester()
 	{
 		final BeanTester beanTester = new BeanTester();
-		beanTester.testBean(LocalDateTimeExtensions.class);
+		beanTester.testBean(ZonedDateTimeExtensions.class);
 	}
+
 }