+++ src/test/java/io/github/astrapi69/time/formatter/DateTimeFormatterFactoryTest.java
@@ -29,8 +29,9 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import java.time.format.DateTimeFormatter;
 
 import org.junit.jupiter.api.Test;
+import org.meanbean.test.BeanTester;
 
-import io.github.astrapi69.time.enums.DatePattern;
+import io.github.astrapi69.time.enumtype.DatePattern;
 
 /**
  * The unit test class for the class {@link DateTimeFormatterFactory}
@@ -38,11 +39,24 @@ import io.github.astrapi69.time.enums.DatePattern;
 class DateTimeFormatterFactoryTest
 {
 
+	/**
+	 * Test method for {@link DateTimeFormatterFactory#newDateTimeFormatter(String)}
+	 */
 	@Test
 	void newDateTimeFormatter()
 	{
 		DateTimeFormatter dateTimeFormatter = DateTimeFormatterFactory
 			.newDateTimeFormatter(DatePattern.DDMM);
 		assertNotNull(dateTimeFormatter);
+	}
+
+	/**
+	 * Test method for {@link DateTimeFormatterFactory}
+	 */
+	@Test
+	public void testWithBeanTester()
+	{
+		final BeanTester beanTester = new BeanTester();
+		beanTester.testBean(DateTimeFormatterFactory.class);
 	}
 }