+++ src/test/java/io/github/astrapi69/time/interval/IntervalExtensionsTest.java
@@ -55,9 +55,6 @@ public class IntervalExtensionsTest
 
 	/**
 	 * Sets up method will be invoked before every unit test method
-	 *
-	 * @throws Exception
-	 *             is thrown if an exception occurs
 	 */
 	@BeforeEach
 	protected void setUp()
@@ -67,9 +64,6 @@ public class IntervalExtensionsTest
 
 	/**
 	 * Tear down method will be invoked after every unit test method
-	 *
-	 * @throws Exception
-	 *             is thrown if an exception occurs
 	 */
 	@AfterEach
 	protected void tearDown()
@@ -208,6 +202,39 @@ public class IntervalExtensionsTest
 		startDate = LocalDate.parse("2007-11-07", formatter).atStartOfDay();
 		endDate = startDate.plus(1, ChronoUnit.MONTHS);
 
+		timeRangeToCheck = Interval.of(startDate.toInstant(ZoneOffset.UTC),
+			endDate.toInstant(ZoneOffset.UTC));
+
+		actual = IntervalExtensions.isBetween(timeRange, timeRangeToCheck);
+		expected = false;
+		assertEquals(actual, expected);
+
+		actual = timeRange.abuts(timeRangeToCheck);
+		expected = false;
+		assertEquals(actual, expected);
+
+		actual = timeRange.encloses(timeRangeToCheck);
+		expected = false;
+		assertEquals(actual, expected);
+
+		actual = timeRange.isAfter(timeRangeToCheck);
+		expected = false;
+		assertEquals(actual, expected);
+
+		actual = timeRange.isBefore(timeRangeToCheck);
+		expected = false;
+		assertEquals(actual, expected);
+
+		// new scenario...
+		startDate = LocalDate.parse("2007-11-08", formatter).atStartOfDay();
+		endDate = startDate.plus(2, ChronoUnit.MONTHS);
+
+		timeRange = Interval.of(startDate.toInstant(ZoneOffset.UTC),
+			endDate.toInstant(ZoneOffset.UTC));
+
+		startDate = LocalDate.parse("2007-11-07", formatter).atStartOfDay();
+		endDate = startDate.plus(10, ChronoUnit.MONTHS);
+		// timeRangeToCheck => 2007-11-07 till 2008-09-07
 		timeRangeToCheck = Interval.of(startDate.toInstant(ZoneOffset.UTC),
 			endDate.toInstant(ZoneOffset.UTC));