+++ src/test/java/io/github/astrapi69/time/query/NextEasternQueryTest.java
@@ -29,6 +29,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.time.LocalDate;
 
 import org.junit.jupiter.api.Test;
+import org.meanbean.test.BeanTester;
 
 /**
  * The unit test class for the class {@link NextEasternQuery}
@@ -67,6 +68,21 @@ public class NextEasternQueryTest
 		actual = input.query(new NextEasternQuery());
 		expected = LocalDate.of(1701, 3, 27);
 		assertEquals(actual, expected);
+
+		input = LocalDate.of(2700, 11, 28);
+		actual = input.query(new NextEasternQuery());
+		expected = LocalDate.of(2701, 4, 21);
+		assertEquals(actual, expected);
+	}
+
+	/**
+	 * Test method for {@link NextEasternQuery}
+	 */
+	@Test
+	public void testWithBeanTester()
+	{
+		final BeanTester beanTester = new BeanTester();
+		beanTester.testBean(NextEasternQuery.class);
 	}
 
 }