+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 ## Change log
 ----------------------
 
+Version 2.1
+-------------
+
+ADDED:
+
+- new module-info.java file with definition of required modules and packages to export
+
+CHANGED:
+
+- update of gradle to new version 6.9.0
+- rename of package enums to enumtype
+
 Version 2
 -------------
 
@@ -70,5 +82,3 @@ CHANGED:
 - update of parent version
 - update of documentation of README.md
 - unit tests extended for improve code coverage
-
-